Refactor path shortening: eliminate circular dependencies, extract strategies (#111)
* Initial plan

* Refactor: Move path shortening logic from writer.py to path_utils.py

- Move update_invalid_paths() and _convert_paths_to_strings() to path_utils.py
- Extract strategy helper functions (attempt_symlink_strategy, etc.)
- Add _is_file_property() and _try_strategies() helpers
- Update imports in writer.py and classes.py
- Fix test logger names to use path_utils instead of writer
- Add comprehensive documentation to test classes

Results:
- writer.py: 638 → 187 lines (70.7% reduction)
- path_utils.py: 240 → 917 lines (standalone, no circular deps)
- All 33 path shortening tests pass
- All 17 writer tests pass
- No circular import errors
